# openapi_summative_project_2.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /publishers:
    get:
      tags:
        - publisher-controller
      operationId: getPublishers
      responses:
        '202':
          description: Accepted
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publisher'
    put:
      tags:
        - publisher-controller
      operationId: updatePublisher
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Publisher'
        required: true
      responses:
        '204':
          description: No Content
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Publisher'
    post:
      tags:
        - publisher-controller
      operationId: addPublisher
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Publisher'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Publisher'
  /books:
    get:
      tags:
        - book-controller
      operationId: getBooks
      responses:
        '202':
          description: Accepted
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
    put:
      tags:
        - book-controller
      operationId: updateBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
        required: true
      responses:
        '204':
          description: No Content
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Book'
    post:
      tags:
        - book-controller
      operationId: addBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Book'
  /authors:
    get:
      tags:
        - author-controller
      operationId: getAllAuthors
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'
    put:
      tags:
        - author-controller
      operationId: putAuthor
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Author'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Author'
    post:
      tags:
        - author-controller
      operationId: addAuthor
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Author'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Author'
  /publishers/{id}:
    get:
      tags:
        - publisher-controller
      operationId: findPublisherById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '202':
          description: Accepted
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Publisher'
    delete:
      tags:
        - publisher-controller
      operationId: deletePublisher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
  /books/{id}:
    get:
      tags:
        - book-controller
      operationId: getBookById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '202':
          description: Accepted
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Book'
    delete:
      tags:
        - book-controller
      operationId: deleteBook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
  /books/author/{id}:
    get:
      tags:
        - book-controller
      operationId: findBookByAuthorId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '202':
          description: Accepted
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
  /authors/{authorId}:
    get:
      tags:
        - author-controller
      operationId: findAuthorById
      parameters:
        - name: authorId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Author'
    delete:
      tags:
        - author-controller
      operationId: DeleteAuthor
      parameters:
        - name: authorId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
components:
  schemas:
    Book:
      type: object
      properties:
        bookId:
          type: integer
          format: int32
        isbn:
          type: string
        publishDate:
          type: string
          format: date
        authorId:
          type: integer
          format: int32
        title:
          type: string
        publisherId:
          type: integer
          format: int32
        price:
          type: number
    Publisher:
      type: object
      properties:
        publisherId:
          type: integer
          format: int32
        name:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        phone:
          type: string
        email:
          type: string
        books:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Book'
    Author:
      type: object
      properties:
        authorId:
          type: integer
          format: int32
        firstName:
          type: string
        lastName:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        phone:
          type: string
        email:
          type: string
        books:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Book'